reorganize pathfinder, goose mind
