+++ src/ideas/creatures/Goose/ideas/Pathfinding/logic/move.ts
